Initial Commit for my .emacs Configuration
This is a bit of a hybrid between what I had before and a new setup,
based on the emacs-prelude package (github.com/bbatsov/emacs-prelude).
Biggest changes are that I use package.el and el-get.el to update and
init external packages.  The odd packages which can't be found in
repositories now live in a folder called vendor/.  I also have a
folder called personal/ -- I'm not quite sure what I'll do with it,
but it seems like a good idea to keep it.  Anything put in here is
automatically sourced.

There are also a bunch of useful functions from the prelude setup in
the elisp/ directory.  I haven't yet gone through them all (or learned
to use them...), but they seem useful to keep.  Keep this in mind to
go through later.