hostile takeover of zenburn-theme
Note: For a while I was getting a funny error: "Invalid theme name `'",
which I thought was a bug in my code.  Actually, it has to do with
ido-completing-read: if I just hit enter, it returns nil instead of the
first item in the list.  So I was trying to set the theme to nil, which
is invalid.  Also explains why the "theme-name" in the error message was
empty.  So this is something to work on...