attempt to fix ido-ubiquitous compatibility issues
hitting RET straight away returns the empty string, rather than the
first item in the list.  apparently this is needed for a lot of
things (?), but it also screws up newer things.

one thing which still isn't working is `load-theme' -- I'd like to make
RET just return the first item in the list