bugfix in prelude-add-watchwords
I had used `regexp-opt' to combine my watchwords into a single regexp.
but that was a mistake, since it uses parenthesized sub-expressions.
I think that threw off font-lock-add-keywords.

now that I type that, it doesn't make sense... major modes work fine
with regexp-opt expressions.  in any case, this seems to be working
now.